Added Main.py
I believe the program should work completly. I was having trouble with it working across multiple monitors but managed to fix that when digging into how the pyscreenshot worked and did it from a lower lever.

Updated the Setup.bat file to have the correct dependacys installed.

The hotkey can be changed by chaning the HOTKEY variable in that format